mapbackend/requirements.txt
django==4.2.6
psycopg[binary]==3.1.12
djangorestframework==3.14.0
uritemplate==4.1.1
PyYAML==6.0.1
djangorestframework-simplejwt==5.3.0
djangorestframework-gis==1.0
